Fix importer

+++ classes/EoN_Importer.php
@@ -321,36 +321,42 @@ class EoN_Importer {
 		global $lilac;
 
 		// Get Link Host ID in XML
-		$link_service_id=$result["Host"];		
-		if($link_service_id!="") {
-			// Get Link Name in XML
-			$link_object=$this->xml_xpath->query("//Host[@id='".$link_service_id."']");
-			$link_service_name=$link_object->item(0)->getElementsByTagName("Name")->item(0)->nodeValue;
-			// Get Link ID in SQL
-			$link_service_created_id=NagiosHostPeer::getByName($link_service_name);
-			$link_service_created_id=$link_service_created_id->getId();
-			$result["Host"]=$link_service_created_id;
+		if(isset($result["Host"])) {
+			$link_service_id=$result["Host"];		
+			if($link_service_id!="") {
+				// Get Link Name in XML
+				$link_object=$this->xml_xpath->query("//Host[@id='".$link_service_id."']");
+				$link_service_name=$link_object->item(0)->getElementsByTagName("Name")->item(0)->nodeValue;
+				// Get Link ID in SQL
+				$link_service_created_id=NagiosHostPeer::getByName($link_service_name);
+				$link_service_created_id=$link_service_created_id->getId();
+				$result["Host"]=$link_service_created_id;
+			}
 		}
 		
 		// Get Link HostTemplate ID in XML
-		$link_service_id=$result["HostTemplate"];		
-		if($link_service_id!="") {
-			// Get Link Name in XML
-			$link_object=$this->xml_xpath->query("//HostTemplate[@id='".$link_service_id."']");
-			$link_service_name=$link_object->item(0)->getElementsByTagName("Name")->item(0)->nodeValue;
-			// Get Link ID in SQL
-			$link_service_created_id=NagiosHostTemplatePeer::getByName($link_service_name);
-			$link_service_created_id=$link_service_created_id->getId();
-			$result["HostTemplate"]=$link_service_created_id;
+		if(isset($result["HostTemplate"])) {
+			$link_service_id=$result["HostTemplate"];		
+			if($link_service_id!="") {
+				// Get Link Name in XML
+				$link_object=$this->xml_xpath->query("//HostTemplate[@id='".$link_service_id."']");
+				$link_service_name=$link_object->item(0)->getElementsByTagName("Name")->item(0)->nodeValue;
+				// Get Link ID in SQL
+				$link_service_created_id=NagiosHostTemplatePeer::getByName($link_service_name);
+				$link_service_created_id=$link_service_created_id->getId();
+				$result["HostTemplate"]=$link_service_created_id;
+			}
 		}
-
+		
 		// Delete existing service
-		$service_id=$lilac->service_exists($name,$result["Host"],$result["HostTemplate"]);
-		if($service_id!=false) {
-			$service = NagiosServicePeer::retrieveByPK($service_id);
-			if($service) {
-				$service->delete();
-				$this->import_msg[]=$this->EoN_Msg_Success("Service","delete","black",$name);
+		if(isset($result["Host"]) && isset($result["HostTemplate"])) {
+			$service_id=$lilac->service_exists($name,$result["Host"],$result["HostTemplate"]);
+			if($service_id!=false) {
+				$service = NagiosServicePeer::retrieveByPK($service_id);
+				if($service) {
+					$service->delete();
+					$this->import_msg[]=$this->EoN_Msg_Success("Service","delete","black",$name);
+				}
 			}
 		}
